#7: refactor + rename functionality
